requirements.txt
autopep8==1.5.7
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
Pillow==8.2.0
progressbar2==3.53.1
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
six==1.16.0
toml==0.10.2
